requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
cryptography==41.0.7
python-multipart==0.0.6

